# [Sub-issue 4/4] Testing: Comprehensive Integration Tests and Performance Validation