… src-tauri/Cargo.lock
… src-tauri/Cargo.toml
… src-tauri/src/ollama_integration_tests.rs
… src-tauri/test_summary.md
… src-tauri/tests/e2e_ollama_tests.rs
… src-tauri/tests/frontend_integration_simplified.rs